+++ injected_feature_brick/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Add feature structure
 - Add creation of abstract repository and implementation repository injected into {{feature}}_cubit
+- Add post mason generation hook to run `build_runner` flutter command